pyproject.toml
[project]
name = "langgraph-tutorial"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "ddgs>=9.11.4",
    "fastmcp>=3.2.0",
    "langchain>=1.2.12",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-mcp-adapters>=0.2.2",
    "langgraph>=1.1.2",
    "python-dotenv>=1.2.2",
]
